% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{cell_type_plot}
\alias{cell_type_plot}
\title{UMAP Visualization of Cells Grouped by Modality}
\usage{
cell_type_plot(umap, rna_cell_size, rna_label, atac_label)
}
\description{
UMAP Visualization of Cells Grouped by Modality
}
