% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{getmetacell_ATAC}
\alias{getmetacell_ATAC}
\title{getmetacell_ATAC}
\usage{
getmetacell_ATAC(object, cluster.name)
}
\arguments{
\item{object}{A SeuratObject}

\item{cluster.name}{The cluster name in the SeuratObject}
}
\value{
A chromatin accessibility matrix, where rows represent cells and columns represent peaks
}
\description{
This function allows us to obtain the chromatin accessibility matrix of metacells
}
