% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{marginal_nls}
\alias{marginal_nls}
\title{marginal_nls}
\usage{
marginal_nls(
  Y,
  X,
  pairwise_par,
  marginal_init_des,
  u,
  index_B_01_des,
  nls_diag,
  loss_f,
  rho,
  n
)
}
\arguments{
\item{Y}{The data matrix Y}

\item{X}{The data matrix X}

\item{pairwise_par}{The parameters obtained from solving the pairwise interactions part in ADMM}

\item{marginal_init_des}{The initial values of the parameters to be solved}

\item{u}{The dual variables updated in the last iteration}

\item{index_B_01_des}{The index matrix of non-zero positions in the coefficient matrix B}

\item{nls_diag}{The custom non-negative least squares solver function that needs to be called}

\item{loss_f}{The custom loss function}

\item{rho}{dual variable parameter}
}
\value{
The solved parameters and the loss function values
}
\description{
This function is used to solve the optimization problem corresponding to the marginal distribution part in ADMM
}
