% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_construct}
\alias{metacell_construct}
\title{Construct Metacells}
\usage{
metacell_construct(object, resolution, graph.name, algorithm = 1)
}
\arguments{
\item{object}{A Seurat object.}

\item{resolution}{Clustering resolution parameter determining the graining level of the resulting meta-cells.}

\item{graph.name}{The name of the graph to use for clustering.}

\item{algorithm}{The clustering algorithm to apply.}
}
\value{
A Seurat object with meta-cell partitioning information.
}
\description{
Perform meta-cell construction by clustering cells within a Seurat object.
}
