% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_matrix_ATAC}
\alias{metacell_matrix_ATAC}
\title{metacell_matrix_ATAC}
\usage{
metacell_matrix_ATAC(object, cluster.name)
}
\arguments{
\item{object}{a}

\item{cluster.name}{a}
}
\value{
a
}
\description{
metacell_matrix_ATAC
}
