% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_matrix_RNA}
\alias{metacell_matrix_RNA}
\title{Compute Metacell-Level Data Matrix}
\usage{
metacell_matrix_RNA(object, cluster.name, size_factor = 10^6)
}
\arguments{
\item{object}{A Seurat object}

\item{cluster.name}{The name of the cluster}

\item{size_factor}{A scaling factor used to normalize the data}
}
\value{
A matrix representing data at the metacell level.
}
\description{
Compute Metacell-Level Data Matrix
}
