% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{nls_diag}
\alias{nls_diag}
\title{nls_diag}
\usage{
nls_diag(Y, X, pairwise_par, marginal_init, u, rho)
}
\arguments{
\item{Y}{a}

\item{X}{a}

\item{pairwise_par}{a}

\item{marginal_init}{a}

\item{u}{a}

\item{rho}{a}
}
\value{
a
}
\description{
nls_diag
}
