% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{pairwise_loss}
\alias{pairwise_loss}
\title{pairwise_loss}
\usage{
pairwise_loss(
  R,
  r,
  B,
  sigma,
  dist_matrix,
  marginal_par,
  u,
  lambda_prior,
  rho,
  w
)
}
\arguments{
\item{R}{a}

\item{r}{a}

\item{B}{a}

\item{sigma}{a}

\item{dist_matrix}{a}

\item{marginal_par}{a}

\item{u}{a}

\item{lambda_prior}{a}

\item{rho}{a}

\item{w}{a}
}
\value{
a
}
\description{
pairwise_loss
}
