% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\docType{package}
\name{SuperMap-package}
\alias{SuperMap}
\alias{SuperMap-package}
\title{SuperMap: Bridging Unpaired Single-Cell Multimodal Data for Integrative Analyses}
\description{
SuperMap is an R package for the integrative analyses of unpaired single-cell multimodal data.
}
\keyword{internal}
