% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{supermap}
\alias{supermap}
\title{Main Function of SuperMAP for Learning Cross-modal Mappings}
\usage{
supermap(
  input_data,
  w = 2,
  rho = 1,
  lambda_prior = 1 * 10^(-4),
  n_iter = 10,
  d0 = 1e+05,
  ncore = 60
)
}
\arguments{
\item{input_data}{The output from the 'gene_peak_distance' function, including both the modality data and the computed gene-peak distance matrix.}

\item{w}{A hyperparameter that balances the contribution of marginal distributions and pairwise interactions. Default is 2.}

\item{rho}{A hyperparameter for the ADMM optimization algorithm. Controls the penalty of the constraint term. Default is 1.}

\item{lambda_prior}{A hyperparameter specifying the strength of the prior knowledge regularization. Default is 1e-4.}

\item{n_iter}{The number of iterations to run the optimization. Default is 10.}

\item{d0}{A hyperparameter defining distance scale, default is 100kb.}

\item{ncore}{Number of CPU cores to use for parallel computation.}
}
\value{
A list containing the estimated cross-modal mappings and convergence information from the optimization.
}
\description{
This function performs the core SuperMAP optimization to learn the cross-modal mappings between modalities.
}
