% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{diagonal_integration}
\alias{diagonal_integration}
\title{Performing diagonal integration}
\usage{
diagonal_integration(
  supermap_data,
  learned_mappings = learned_mappings,
  rna,
  atac,
  rna_label
)
}
\description{
Performing diagonal integration
}
