% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{diagonal_integration}
\alias{diagonal_integration}
\title{diagonal_integration}
\usage{
diagonal_integration(
  supermap_data,
  learned_mappings = learned_mappings,
  rna,
  atac,
  rna_label
)
}
\arguments{
\item{supermap_data}{a}

\item{learned_mappings}{a}

\item{rna}{a}

\item{atac}{a}

\item{rna_label}{a}
}
\value{
a
}
\description{
diagonal_integration
}
