% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{diagonal_integration}
\alias{diagonal_integration}
\title{Performing diagonal integration}
\usage{
diagonal_integration(rna_data, atac_imputation, rna_label)
}
\description{
Performing diagonal integration
}
