% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{gene_peak_distance}
\alias{gene_peak_distance}
\title{gene_peak_distance}
\usage{
gene_peak_distance(rna, atac, gene_annotation)
}
\arguments{
\item{rna}{The RNA modality data matrix, where column is feature and row is cell.}

\item{atac}{The ATAC modality data matrix, where column is feature and row is cell.}

\item{gene_annotation}{A data frame that contains gene location informance}
}
\value{
A list that contains modality data and peak-gene distance
}
\description{
gene_peak_distance
}
