% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{imputation}
\alias{imputation}
\title{imputation}
\usage{
imputation(supermap_data, learned_mappings, atac)
}
\arguments{
\item{supermap_data}{a}

\item{learned_mappings}{a}

\item{atac}{a}
}
\value{
a
}
\description{
imputation
}
