% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{marginal_nls}
\alias{marginal_nls}
\title{marginal_nls}
\usage{
marginal_nls(
  Y,
  X,
  pairwise_par,
  marginal_init_des,
  u,
  index_B_01_des,
  nls_diag,
  loss_f,
  rho,
  n
)
}
\arguments{
\item{Y}{a}

\item{X}{a}

\item{pairwise_par}{a}

\item{marginal_init_des}{a}

\item{u}{a}

\item{index_B_01_des}{a}

\item{nls_diag}{a}

\item{loss_f}{a}

\item{rho}{a}

\item{n}{a}
}
\value{
a
}
\description{
marginal_nls
}
