% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{marginal_nls}
\alias{marginal_nls}
\title{Marginal distribution term}
\usage{
marginal_nls(
  Y,
  X,
  pairwise_par,
  marginal_init_des,
  u,
  index_B_01_des,
  nls_diag,
  loss_f,
  rho,
  n
)
}
\description{
Marginal distribution term
}
