% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_construct}
\alias{metacell_construct}
\title{metacell_construct}
\usage{
metacell_construct(object, resolution, graph.name, algorithm = 1)
}
\arguments{
\item{object}{a}

\item{resolution}{a}

\item{graph.name}{a}

\item{algorithm}{a}
}
\value{
a
}
\description{
metacell_construct
}
