% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_construct}
\alias{metacell_construct}
\title{Construct Metacells}
\usage{
metacell_construct(object, resolution, graph.name, algorithm = 1)
}
\arguments{
\item{object}{A Seurat object}

\item{resolution}{Clustering resolution parameter controlling the granularity of metacell}

\item{graph.name}{Name of the graph to use for clustering}

\item{algorithm}{Clustering algorithm to use}
}
\value{
A Seurat object with meta-cell partitioning information
}
\description{
Construct Metacells
}
