% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_matrix_ATAC}
\alias{metacell_matrix_ATAC}
\title{Compute Metacell-Level Data Matrix}
\usage{
metacell_matrix_ATAC(object, cluster.name)
}
\arguments{
\item{object}{A Seurat object}

\item{cluster.name}{The name of the cluster}
}
\value{
A matrix representing data at the metacell level.
}
\description{
Compute Metacell-Level Data Matrix
}
