% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_matrix_ATAC}
\alias{metacell_matrix_ATAC}
\title{Compute Meta-cell Level Data Matrix}
\usage{
metacell_matrix_ATAC(object, cluster.name)
}
\arguments{
\item{object}{A Seurat object.}

\item{cluster.name}{The name of the clustering column in the Seurat object's metadata that defines the meta-cell partitions.}
}
\value{
A matrix representing meta-cell level data.
}
\description{
Compute Meta-cell Level Data Matrix
}
