% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_matrix_RNA}
\alias{metacell_matrix_RNA}
\title{metacell_matrix_RNA}
\usage{
metacell_matrix_RNA(object, cluster.name, size_factor = 10^6)
}
\arguments{
\item{object}{a}

\item{cluster.name}{a}

\item{size_factor}{a}
}
\value{
a
}
\description{
metacell_matrix_RNA
}
