% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{metacell_matrix_RNA}
\alias{metacell_matrix_RNA}
\title{Compute Meta-cell Level Data Matrix}
\usage{
metacell_matrix_RNA(object, cluster.name, size_factor = 10^6)
}
\arguments{
\item{object}{A Seurat object.}

\item{cluster.name}{The name of the clustering column in the Seurat object's metadata that defines the meta-cell partitions.}

\item{size_factor}{A scaling factor used to normalize the aggregated data.}
}
\value{
A matrix representing meta-cell level data.
}
\description{
Compute Meta-cell Level Data Matrix
}
