% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{nls_diag}
\alias{nls_diag}
\title{nls_diag}
\usage{
nls_diag(Y, X, pairwise_par, marginal_init, u, rho)
}
\description{
This function utilizes the Levenberg-Marquardt (LM) method to solve the nonlinear least squares problem
}
