% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{pairwise_loss}
\alias{pairwise_loss}
\title{Computing pairwise loss}
\usage{
pairwise_loss(
  R,
  r,
  B,
  sigma,
  dist_matrix,
  marginal_par,
  u,
  lambda_prior,
  rho,
  w
)
}
\description{
Computing pairwise loss
}
