% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{pairwise_optim}
\alias{pairwise_optim}
\title{Pairwise interactions term}
\usage{
pairwise_optim(
  Y,
  X,
  B_initial,
  sigma_init,
  index_B_01,
  dist_matrix,
  lambda_B,
  rho,
  w,
  marginal_par,
  u,
  iteration,
  d0
)
}
\description{
Pairwise interactions term
}
