% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{pairwise_optim}
\alias{pairwise_optim}
\title{pairwise_optim}
\usage{
pairwise_optim(
  Y,
  X,
  B_initial,
  sigma_init,
  index_B_01,
  dist_matrix,
  lambda_B,
  rho,
  w,
  marginal_par,
  u,
  iteration,
  d0
)
}
\arguments{
\item{Y}{The data matrix Y}

\item{X}{The data matrix X}

\item{B_initial}{The initial value of the coefficient matrix B}

\item{sigma_init}{The initial value of the variance sigma}

\item{index_B_01}{The index matrix of non-zero positions in the coefficient matrix B}

\item{dist_matrix}{prior-related weight matrix D}

\item{lambda_B}{prior penalty parameter}

\item{rho}{dual variable parameter}

\item{w}{parameter for weighting marginal distributions and pairwise interactions}

\item{marginal_par}{The parameters obtained from solving the marginal distribution part in ADMM}

\item{u}{The dual variables updated in the last iteration}

\item{iteration}{The number of iterations}

\item{d0}{The hyperparameter d0}
}
\value{
Return the solved parameters B, sigma, and the loss value for each iteration.
}
\description{
This function is used to solve the optimization problem corresponding to the pairwise interactions part in ADMM
}
