% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{supermap}
\alias{supermap}
\title{supermap}
\usage{
supermap(input_data, w, rho, lambda_prior, n_iter = 10, d0 = 1e+05, ncore = 60)
}
\arguments{
\item{input_data}{a}

\item{w}{a}

\item{rho}{a}

\item{lambda_prior}{a}

\item{n_iter}{a}

\item{d0}{a}

\item{ncore}{a}
}
\value{
a
}
\description{
supermap
}
