% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermap.R
\name{supermap}
\alias{supermap}
\title{supermap}
\usage{
supermap(
  Y_unpaired,
  X_unpaired,
  index_B_nonzero,
  B_dist,
  w,
  rho,
  lambda_prior,
  n_iter = 20,
  d0 = 1e+05
)
}
\arguments{
\item{Y_unpaired}{The data matrix Y}

\item{X_unpaired}{The data matrix X}

\item{index_B_nonzero}{The index matrix of non-zero positions in the coefficient matrix B}

\item{B_dist}{prior-related weight matrix D}

\item{w}{parameter for weighting marginal distributions and pairwise interactions}

\item{rho}{dual variable parameter}

\item{lambda_prior}{prior penalty parameter}

\item{n_iter}{The number of iterations}

\item{d0}{The hyperparameter d0, with a default value of 100,000}
}
\value{
The estimated model parameters, as well as the loss value for each iteration
}
\description{
This function utilizes the ADMM algorithm to estimate parameters in SuperMap model.
}
